refactor: PageResponse分页字段从long改为int，与PageResult保持一致

+++ backend/src/main/java/com/linyi/pig/common/model/PageResponse.java
@@ -8,12 +8,12 @@ public class PageResponse {
     /**
      * 当前页码（从1开始）
      */
-    private long pageNum;
+    private int pageNum;
 
     /**
      * 每页记录数
      */
-    private long pageSize;
+    private int pageSize;
 
     /**
      * 排序字段
